Protocols: Set `clean_workdir` default to `False` (#919)
The current default in the protocol of all work chains is to set
`clean_workdir` to `True`. This can several use cases where users have
to restart from previous calculations impossible when the default
inputs are used, which can be particularly confusing and frustrating for
beginners.

Moreover, cleaning the working directory removes potentially valuable
data that is not retrieved by the plugin (e.g. the charge density).
Conversely, _not_ cleaning the working directory only really is a
problem in case the user is running _a lot_ of calculations, and can be
fixed easily by manually cleaning the working directories.

Here we set the default value of `clean_workdir` to `False` for all work
chain protocols.