`PwCalculation`: Add exit code and handler for `scale_h` error (#913)
This exit code appears when a cell relaxation produces a significant
volume contraction. This means the pseudopotentials tables must be
recalculated. This parameter is controlled by `CELL.cell_factor`.

The solution, as suggested by the Quantum ESPRESSO error itself, is to
restart with an increased `cell_factor`. The handler restarts from
scratch using the last output structure, doubling the cell factor which
should be 2 by default.