Protocols: consider `pbc` of `StructureData` (#907)
Currently none of the `get_builder_from_protocol()` methods consider the
periodic boundary conditions (`pbc`) of the input `StructureData`. Here
we add support for several cases to the `PwBaseWorkChain` and
`PwRelaxWorkChain`:

* 2D structures in the x-y plane with periodic boundary conditions for
  the first two lattice vectors (`pbc == [True, True, False]`). For this
  case, the `PwBaseWorkChain` protocol will set `SYSTEM.assume_isolated`
  to `2D`, and the `PwRelaxWorkChain` protocol will use
  `CELL.do_free = '2Dxy'` in case the `RelaxType` is `CELL` or
  `POSITIONS_CELL`.
* 1D structures defined on an arbitrary lattice vector will use the
  corresponding `CELL.do_free` setting (`'x'`, `'y'`, `'z'`) if
  `RelaxType` is `CELL` or `POSITIONS_CELL`.

So far, no validation is done to see if the structure has the correct
`cell` specified. This could potentially be added to the `PwCalculation`
validation.

Co-authored-by: Marnik Bercx <mbercx@gmail.com>