`BasePwCpInputGenerator`: Remove superfluous validation
In the `prepare_for_submission` script of the `BasePwCpInputGenerator`,
there still is some validation for the pseudopotentials (see that the elements
match those in the provided `structure`) as well as the `FIXED_COORDS`
setting. 

In general, we want to do all validation using validators of the appropriate port
or port name space. In the case of these two inputs, however, this validation is
already done by proper validators, so we can simply remove it.

Even in case the `structure` port is not in the name space because a wrapping
work chain excluded it when exposing the inputs, the validation will still occur
when the process is submitted in the steps of the work chain.