`PwCalculation`: refactor `parent_folder` validation
In 3fea26b we added validation for restarting from a `parent_folder` to the 
`PwCalculation`. Here we ran into the following problem:

1. We want to check that a `parent_folder` is present when `CONTROL.calculation`
   is `nscf` or `bands`
2. The `parent_folder` will not be present in the initial inputs of e.g. the
   `PwBandsWorkChain` `bands` name space, since it is only created on runtime.
3. The validation will hence always fail, raising a warning when the builder is
   put together using the `get_builder_from_protocol()` method.

Our solution at the time was to split up the validation in a base and full one,
and override the validator. However, a more elegant solution is to check if the 
`parent_folder` is in the name space, and only do the restart validation in this
case. Wrapping work chains that dynamically assign this input should exclude the
`parent_folder` port when exposing the inputs of the `PwBaseWorkChain`.