Devops: Update `tox` configuration
The `tox` configuration was still using outdated Python versions. Here the
Python versions are updated to those supported by the package, the environments
are cleaned up a bit, and the `pre-commit` environment is added to those run by
default.

The docs environment are removed, since these aren't usually used and we still
have an issue with the `intersphinx_mapping` to `aiida-core`.