Docs: switch to `sphinx-book-theme` from `pydata_sphinx_theme`
The main reason for the switch is to have access to theme-specific elements like
margin notes. Another major change compared to `pydata_sphinx_theme` is the
absence of a top-level header, which has us move the contents overview into the
sidebar. This has the advantage that the user immediately has a bird's eye
glance of the documentation contents. Until the documentation becomes larger,
this approach is preferable.

Several changes are made to the sidebar:

1. Override the default `icon-links.html` component to include an AiiDA icon
   that links to the AiiDA website.
2. Override the `sbt-sidebar-nav.thml` component to adapt the default maxdepth.
3. A logo is added for dark mode.

Some other additions and tweaks:

1. Added badged to the landing page instead of the release version.
2. Adapt the sidebar `toctree` to include captions and point directly to some
   subsections.
3. The Python API now has a `toctree` element with a clear name, and the
   `maxdepth` is adapted to make it more readable.
4. Several panels on the landing page are removed, since the corresponding
   top-level sections are no longer part of the `toctree`. This also makes the
   remaining panels to the `Get started` and tutorial sections more prominent.

Finally, the `intersphinx_mapping` to the `aiida-core` docs is fixed, along with
several broken references, mostly in the Hubbard docstring.

The `pydantic` and `typing` related intersphinx warnings are ignored using
`nitpick_ignore_regex`, since pydantic doesn't use `sphinx` and there seem to
be issues with referencing `typing` classes.