`PwBaseWorkChain`: Remove disabling of resource validation
The `metadata.options.resources` port, exposed from `PwCalculation` was
set to not required, which disables its validation. This is problematic
when a parent workflow is launched with incorrect resources provided for
the `PwCalculation` subprocess. In this case the problem will only be
noticed when the `PwCalculation` gets actually launched, at which point
the parent process has already been created in the database, leaving it
excepted.

The disabling is a remnant from the time that the `PwBaseWorkChain`
provided an `automatic_parallelization` option, which would determine
the optimal resources for the `PwCalculation` automatically, but this
feature has long been removed. Therefore the resource validation can be
reenabled again.