`Q2rCalculation`: Add the `output_parameters` output to the spec
The `Q2rCalculation` did not declare the `output_parameters` output even
though the `Q2rParser` attaches it always. This would lead to the
calculation excepting because an unknown output would be attached. This
was uncaught by the test because the `Parser.parse_from_node` method
does not automatically validate the outputs. This is because the output
spec that is checked is that of the `calcfunction` (which accepts
anything) and not that of the `Q2rCalculation`.