Dependencies: Update `pydantic~=2.4`
The requirement for `pydantic` was pinned to `v1` since `v2` has a lot
of backwards incompatible changes and it is difficult to provide a
version that is compatible with both versions.

As of `v2.5.0`, `aiida-core` also directly depends on `pydantic` and it
requires `~=2.4`, so here we apply the same requirement. The deprecated
code is replaced.

As a side effect, a test for `PwCalculation` started failing. It was
testing that no warnings were raised for specific inputs, but some
warnings _were_ being raised. These were not the warnings tested for in
the tests though, but raised by SQLAlchemy. AiiDA v2.5 upgraded to
SQLAlchemy v2 which as of v2.0.19 started emitting a warning. This is
ignored by a filter in `aiida-core`, but this is made undone by `pytest`.
The warning is filtered again in the `pyproject.toml` config for
`pytest` but this is not considered by `pytest.warns`. Therefore, the
warnings in `PwCalculation` are turned into the more specific
`UserWarning` such that the test can explicitly check for those.