Devops: Address various deprecation warnings from `aiida-core`
AiiDA v2.5 upgraded to Sqlalchemy v2 which as of v2.0.19 emits a warning
that is entirely consequential but was too difficult to actually address.
As a workaround, `aiida-core` simply filters the warning. This works for
normal usage, but `pytest` eliminates the filter and so the warnings do
show up in the final warning report of `pytest`. Therefore the warning
has to be explicitly ignored in the `pytest` config as well.