Test/Improve `PhBaseWorkChain` overrides/protocol
The `PhBaseWorkChain.get_builder_from_protocol` was untested.
Tests are added, and the `electronic_type` input is added to allow
for the computation of the dielectric and effective charge tensors
when an insulator is in input. This is useful in other workchains, so
that the `electronic_type` input can be passed seamlessly as a
kwargs input at an higher level.