👌 `PhBaseWorkChain`: better handler for convergence
Simply reducing the `alpha_mix` does not always fix convergence issues. Here we
increase the default number of iterations into the mixing by setting `nmix_ph` to 8 in
the `PhBaseWorkChain` protocol, and check this value when handling the error. The rate
of the `alpha_mix` reduction is also increased, and below a certain threshold a multi
mix approach is tried. Afterwards, the handler finishes the strategies.