`PhBaseWorkChain`: allow generation of q-point mesh via `qpoints_distance`
Allow the user to specify the density of the q-point mesh using a new `qpoint_distance`
input on the `PhBaseWorkchain`. The `qpoints` input of the wrapper `PhCalculation` is
also moved replicate the usage of the `PwBaseWorkChain` for specifying Brillouin zone
sampling.

The `set_qpoints` step is added to the outline, where either:

1. the provided `qpoints` are directly passed to the inputs
2. the structure is retrieved via the `creator` of the `parent_folder` input, and
   subsequently used in combination with the `qpoints_distance` to generate the q-points
   mesh.

A validator and basic protocol is added, along with the `qpoints_force_parity` input,
which can be used to guarantee that the q-point mesh has an even number along any
periodic direction of the lattice.