`PdosWorkChain`: Fix constrained magnetization case
When using constrained magnetization via the `SYSTEM.tot_magnetization` input of `pw.x`,
the calculation outputs two different fermi levels for the two spin channels:
`fermi_energy_up` and `fermi_energy_down`. In this case the `PdosWorkChain` would except
since it extracts the `fermi_energy` output in the `inspect_nscf` step of the outline.

Here the step will check if the `fermi_energy` value is in the output, and look for the
two values of the up/down spin channels in case it isn't. The fermi level of the NSCF
step is then set to the maximum of the two levels.

The `fixture_code` fixture is updated to use a query instead of `load_code`, since
using `pymark.parametrize` with the hierarchy of fixtures can apparently create multiple
instances of codes with the same label in the testing profile.

Co-authored-by: Marnik Bercx <mbercx@gmail.com>