Docs: Remove `aiida.manage.configuration.load_documentation_profile` (#1000)
This was necessary when Django backends were still supported, but since
its removal in AiiDA v2.0 just loading the ORM no longer actually needs
to connect to a database. In principle, the documentation no longer
needs a profile whatsoever, except for `aiida-quantumespresso` the
AutoAPI still needs it because in loading the specifications of certain
`Process` classes, a profile is loaded. So the docs `conf.py` now loads
an empty dummy profile for this purpose.