`PwBaseWorkChain`: new handler for BFGS history failure
When the BFGS history fails during ionic minimizaiton,
the current handler simply restarts from scratch. This 
simplistic solution has limited effectiveness. We improve
by trying first to lower the trusted radius. If the history
fails again, then we switch to the `damp` algorithm. The 
latter solution proved to converge the ionic minimization 
after the occurrence of a BFGS history failure on a large
test set.

Co-authored-by: Marnik Bercx <mbercx@gmail.com>