Improve inputs and refold with thr
The refolding wasn't folding numbers
close numerically to 1, e.g. 0.99999999.
To circumvent the issue, a threshold is set,
and it is possible to control it via input.
Also, now the inputs for nearest neighbours
are not passed via kwargs, but via an explicit
dictionary input, which is then given via kwargs
to the NN class.