`HubbardStructureData`: Add validation on site indices
The `append_hubbard_parameter()` method of the `HubbardStructureData` class currently
does not verify that the `atom_index` and `neighbour_index` values correspond to valid
site indices in the structure. This means that the method will:

- fail in case `translation` is set to `None` (as it attempts to access these site indices to obtain
  the distance and image of the nearest neighbour using periodic boundary conditions)
- succeed in case the translation is specified, but fail once the code tries to write Hubbard card
  during the `prepare_for_submission` method of the `PwCalculation`.

To have more consistent behaviour and avoid exceptions during the `upload` step of the
submission process, we check here that the `atom_index` and `neighbour_index` are valid and raise
a `ValueError` if they are not.