`HubbardStructureData`: add compatibility for <3D structures
The `HubbardStructureData.append_hubbard_parameter()` method currently fails for
structures that don't have 3-dimensional periodic boundary conditions. This is caused
by a restriction in the StructureData.get_pymatgen_structure()` method, which raises a
`ValueError` if the periodic boundary conditions are not (True, True, True). This has
been fixed in a recent commit on `aiida-core`:

https://github.com/aiidateam/aiida-core/commit/adcce4bcd0b59c8371be73058a060bedcaba40f6

But this will take some time to be released and would restrict our compatibility to
`aiida-core>=2.6.0`.

Here we directly construct a list of `PeriodicSite` objects from the `sites` list of the
object obtained from `StructureData.get_pymatgen()`, which can be both either a
pymatgen `Structure` or `Molecule`, depending on the dimensionality of the structure.

Co-authored-by: AndresOrtegaGuerrero <34098967+AndresOrtegaGuerrero@users.noreply.github.com>