`PhBaseWorkChain`: skip q-points validation if ports are excluded
Currently, the top-level `validate_inputs` validator for the `PhBaseWorkChain` will
always check if either the `qpoints` or `qpoints_distance` inputs are provided. However,
there are cases where a higher-level work chain that wraps the `PhBaseWorkChain` might
provide the q-points input on the fly. Such a work chain would have to set the
`PhBaseWorkChain` validator to `None` in its own `define` method.

However, as also discussed for a similar case for the `PwCalculation` in a389629,
it's more elegant to check if the `qpoints_distance` and `qpoints` ports are present in
the namespace. Typically a work chain that provides the q-points input on the fly will
exclude these ports when exposing the `PhBaseWorkChain` inputs. By checking if the ports
are present in the validator, we avoid having to set the validator to `None` as well in
the higher-level work chain.