`PpCalculation`: Make parsing of output files optional (#1029)
The `parse_data_files` option is added. When switched to `False` the
parser will not parse the outputs files but just keep the raw files. The
existing option `keep_plot_file` is deprecated in favor of the renamed
`keep_data_files` option to make it coherent with the new option.