👌 `PwBaseWorkChain`: Always do full restart for `ERROR_OUT_OF_WALLTIME`
The current error handler for the `ERROR_OUT_OF_WALLTIME` exit code of the
`PwCalculation` will restart from scratch in case the structure has changed during the
`pw.x` run, as is typically the case for `relax`/`vc-relax` calculations. For larger
structures and more complex calculations - such as those using Hubbard corrections -
this can be quite inefficient since obtaining the electronic ground state is often
more challenging and hence expensive.

Here we adapt the error handler to always do a full restart from the previous
calculation. In case the structure has changed, we still set it as the input structure
of the restart calculation. Even though Quantum ESPRESSO will restart from the `.save`
directory and hence ignore the structure information in the input file, this at least
makes the structure in the provenance consistent with the one run by Quantum ESPRESSO.