📚 `README.md`: Correct `aiida-core` compatibility
In 740e0bec0e68b3229367b9f10b181a925616c08b support was added for `pydantic~=2.4`. This
effectively changed the compatible `aiida-core` version to v2.5 or above, since from
this version `aiida-core` also depends directly on `pydantic~=2.4`.

However, we never adapted the compatibility matrix in the `README.md` to reflect this
restriction. This commit updates the compatibility matrix to reflect the correct
compatible versions of `aiida-core`.