🐛 `ProjwfcParser`: Fix DOS parsing for spin-polarised case
Currently, the `ProjwfcParser` does not correctly parse the total DOS data in the case
of spin-polarised calculations. Only the first column is parsed, which corresponds to
spin-up density only. This issue passed the tests unnoticed since the regression tests
only check the attributes of the `Dos` output, not the actual content.

Here we adapt the `ProjwfcParser` to treat the spin-polarised case `nspin = 2` case
correctly, constructing an `XyData` node that has two `y` arrays, one for each spin.
The tests for the spin-polarised case are also adapted to add the `XyData` contents
to the regression files.