Fix to allow retrieval of nbands params from SCF when no SCF is to be launched by PdosWorkChain
In the case where no SCF workchain is launched, `workchain_scf` is not in `self.ctx` so an error occurs when trying to get the `output_parameters` of the SCF workchain and retrieve the number of bands used.
To fix this: `workchain_scf` is added to `self.ctx` when 'scf' is not in `self.inputs` by getting the caller of the creator of the `parent_folder` specified in 'nscf' inputs.